More tweaks to train limit math
Still some issues with multiple trains.  Usually because a non-dispatched train from the same surface grabs a reservation while a dispatched train is arriving.  This should be fixable if we can make it so that trains on the same surface always have to be dispatched by the mod too.

Next tasks:
- Redesign the scheduler to look for the closest train to each stop rather than the closest stop for the waiting trains in order.
- Make it so that trains from the same surface get temporary stops too.