Dispatcher priority works!
- Rewrote the dispatcher to prefer trains on the nearest surface.
- Surface link table now includes link to itself with cost 0 and no schedule
- Added some code to try and detect when a train gets double-counted between arrivals and vanilla reservations.

TODO: Trains which dispatch themselves need to be interrupted and have temporary stops added.  This is so that the stop limit does not have to account for them pathing a long distance to the actual stop.  That way, the only time the stop limit needs to be >0 is when a train is just arriving.  All other times, we use the mod dispatcher to decide when trains go.